Revert "Add "firebase functions:shell" for emulation of all functions types (#428)"
This reverts commit b7237c4b69a827943dab7156eaeb610a6d63aa0d.
