+++ package.json
@@ -98,6 +98,6 @@
     "sinon-chai": "^2.8.0"
   },
   "optionalDependencies": {
-    "@google-cloud/functions-emulator": "^1.0.0-alpha.23"
+    "@google-cloud/functions-emulator": "^1.0.0-alpha.22"
   }
 }
